// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ProjectStatus = {
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
  ON_HOLD: 'ON_HOLD',
  CANCELLED: 'CANCELLED'
} as const

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus]


export const TaskStatus = {
  TODO: 'TODO',
  IN_PROGRESS: 'IN_PROGRESS',
  DONE: 'DONE'
} as const

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus]


export const Priority = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH'
} as const

export type Priority = (typeof Priority)[keyof typeof Priority]


export const QuoteStatus = {
  DRAFT: 'DRAFT',
  SENT: 'SENT',
  ACCEPTED: 'ACCEPTED',
  REJECTED: 'REJECTED'
} as const

export type QuoteStatus = (typeof QuoteStatus)[keyof typeof QuoteStatus]


export const PostStatus = {
  DRAFT: 'DRAFT',
  SCHEDULED: 'SCHEDULED',
  PUBLISHED: 'PUBLISHED'
} as const

export type PostStatus = (typeof PostStatus)[keyof typeof PostStatus]


export const EventType = {
  RDV_TEL: 'RDV_TEL',
  PERSONNEL: 'PERSONNEL',
  DEADLINE: 'DEADLINE'
} as const

export type EventType = (typeof EventType)[keyof typeof EventType]